refacto(archi): move Blog
