+++ src/Blog/Domain/PostSeo.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Domain\Blog;
+namespace App\Blog\Domain;
 
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
