+++ src/Http/Controller/Admin/DashboardController.php
@@ -2,13 +2,13 @@
 
 namespace App\Http\Controller\Admin;
 
-use App\Domain\Blog as Blog;
+use App\Blog as Blog;
 use App\Domain\CV as CV;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
 use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 class DashboardController extends AbstractDashboardController
 {
@@ -32,7 +32,7 @@ class DashboardController extends AbstractDashboardController
         yield MenuItem::linkToDashboard('Dashboard', 'fa fa-home');
 
         yield MenuItem::section('Blog');
-        yield MenuItem::linkToCrud('Posts', 'fas fa-newspaper', Blog\Post::class);
+        yield MenuItem::linkToCrud('Posts', 'fas fa-newspaper', Blog\Domain\Post::class);
 
         yield MenuItem::section('CV');
         yield MenuItem::linkToCrud('Professional experiences', 'fas fa-briefcase', CV\ProfessionalExperience::class);
